+++ addon/mixins/application-route-mixin.js
@@ -1,7 +1,11 @@
+import Mixin from '@ember/object/mixin';
+import { A } from '@ember/array';
+import { computed } from '@ember/object';
+import { getOwner } from '@ember/application';
+import { inject } from '@ember/service';
 import Ember from 'ember';
 import Configuration from './../configuration';
-
-const { inject, Mixin, A, run: { bind }, testing, computed, getOwner } = Ember;
+import isFastBoot from 'ember-simple-auth/utils/is-fastboot';
 
 /**
   The mixin for the application route, __defining methods that are called when
@@ -50,13 +54,9 @@ export default Mixin.create({
     @type SessionService
     @public
   */
-  session: inject.service('session'),
+  session: inject('session'),
 
-  _isFastBoot: computed(function() {
-    const fastboot = getOwner(this).lookup('service:fastboot');
-
-    return fastboot ? fastboot.get('isFastBoot') : false;
-  }),
+  _isFastBoot: isFastBoot(),
 
   /**
     The route to transition to after successful authentication.
@@ -80,9 +80,7 @@ export default Mixin.create({
       ['authenticationSucceeded', 'sessionAuthenticated'],
       ['invalidationSucceeded', 'sessionInvalidated']
     ]).forEach(([event, method]) => {
-      this.get('session').on(event, bind(this, () => {
-        this[method](...arguments);
-      }));
+      this.get('session').on(event, (...args) => this[method](...args));
     });
   },
 
@@ -133,11 +131,11 @@ export default Mixin.create({
     @public
   */
   sessionInvalidated() {
-    if (!testing) {
+    if (!Ember.testing) {
       if (this.get('_isFastBoot')) {
-        this.transitionTo(Configuration.baseURL);
+        this.transitionTo(Configuration.rootURL);
       } else {
-        window.location.replace(Configuration.baseURL);
+        window.location.replace(Configuration.rootURL);
       }
     }
   }