+++ config/environment.js
@@ -1,8 +1,5 @@
 'use strict';
 
-/* eslint-env node */
-/* eslint-disable no-var, object-shorthand */
-
 module.exports = function(/* environment, appConfig */) {
   return { };
 };