+++ tests/dummy/app/components/login-form.js
@@ -1,7 +1,6 @@
-import Ember from 'ember';
+import { inject as service } from '@ember/service';
+import Component from '@ember/component';
 import config from '../config/environment';
-
-const { inject: { service }, Component } = Ember;
 
 export default Component.extend({
   session: service('session'),
@@ -9,9 +8,13 @@ export default Component.extend({
   actions: {
     authenticateWithOAuth2() {
       let { identification, password } = this.getProperties('identification', 'password');
-      this.get('session').authenticate('authenticator:oauth2', identification, password).catch((reason) => {
-        this.set('errorMessage', reason.error);
-      });
+      this.get('session').authenticate('authenticator:oauth2', identification, password)
+        .then(() => {
+          this.get('rememberMe') && this.set('session.store.cookieExpirationTime', 60 * 60 * 24 * 14);
+        })
+        .catch((reason) => {
+          this.set('errorMessage', reason.error);
+        });
     },
 
     authenticateWithFacebook() {