+++ tests/unit/authenticators/torii-test.js
@@ -1,10 +1,8 @@
-import Ember from 'ember';
+import RSVP from 'rsvp';
 import { describe, beforeEach, it } from 'mocha';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import Torii from 'ember-simple-auth/authenticators/torii';
-
-const { RSVP } = Ember;
 
 describe('ToriiAuthenticator', () => {
   let authenticator;
@@ -22,9 +20,14 @@ describe('ToriiAuthenticator', () => {
   describe('#restore', function() {
     function itDoesNotRestore(data) {
       it('returns a rejecting promise', function() {
-        return authenticator.restore(data).catch(() => {
-          expect(true).to.be.true;
-        });
+        return authenticator.restore(data).then(
+          () => {
+            expect(false).to.be.true;
+          },
+          () => {
+            expect(true).to.be.true;
+          }
+        );
       });
     }
 
@@ -42,9 +45,9 @@ describe('ToriiAuthenticator', () => {
           sinon.stub(torii, 'fetch').returns(RSVP.resolve({ some: 'other data' }));
         });
 
-        it('returns a promise that resolves with the session data', function() {
-          return authenticator.restore({ some: 'data', provider: 'provider' }).then((data) => {
-            expect(data).to.eql({ some: 'other data', provider: 'provider' });
+        it('returns a promise that resolves with the session data merged with the data fetched from torri', function() {
+          return authenticator.restore({ some: 'data', provider: 'provider', another: 'prop' }).then((data) => {
+            expect(data).to.eql({ some: 'other data', provider: 'provider', another: 'prop' });
           });
         });
       });