+++ tests/unit/initializers/setup-session-test.js
@@ -1,5 +1,10 @@
 import Ember from 'ember';
-import { describe, beforeEach, afterEach, it } from 'mocha';
+import {
+  describe,
+  beforeEach,
+  afterEach,
+  it
+} from 'mocha';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import setupSession from 'ember-simple-auth/initializers/setup-session';
@@ -14,6 +19,7 @@ describe('setupSession', () => {
       register() {},
       injection() {}
     };
+    Ember.testing = true;
   });
 
   it('registers the session', function() {