+++ tests/unit/mixins/authenticated-route-mixin-test.js
@@ -1,4 +1,6 @@
-import Ember from 'ember';
+import Mixin from '@ember/object/mixin';
+import RSVP from 'rsvp';
+import Route from '@ember/routing/route';
 import { describe, beforeEach, it } from 'mocha';
 import { expect } from 'chai';
 import sinon from 'sinon';
@@ -8,8 +10,6 @@ import Configuration from 'ember-simple-auth/configuration';
 import EphemeralStore from 'ember-simple-auth/session-stores/ephemeral';
 
 import createWithContainer from '../../helpers/create-with-container';
-
-const { Mixin, RSVP, Route } = Ember;
 
 describe('AuthenticatedRouteMixin', () => {
   let route;
@@ -46,10 +46,10 @@ describe('AuthenticatedRouteMixin', () => {
 
       containerMock.lookup.withArgs('service:cookies').returns(cookiesMock);
       containerMock.lookup.withArgs('service:fastboot').returns(fastbootMock);
+      containerMock.lookup.withArgs('service:session').returns(session);
 
       route = createWithContainer(Route.extend(MixinImplementingBeforeModel, AuthenticatedRouteMixin, {
         // pretend this is never FastBoot
-        _isFastBoot: false,
         // replace actual transitionTo as the router isn't set up etc.
         transitionTo() {}
       }), { session }, containerMock);
@@ -91,12 +91,9 @@ describe('AuthenticatedRouteMixin', () => {
 
       it('sets the redirectTarget cookie in fastboot', function() {
         fastbootMock.get.withArgs('request.protocol').returns('https');
+        fastbootMock.get.withArgs('isFastBoot').returns(true);
 
         let cookieName = 'ember_simple_auth-redirectTarget';
-
-        route.reopen({
-          _isFastBoot: true
-        });
 
         route.beforeModel(transition);
         expect(cookiesMock.write).to.have.been.calledWith(cookieName, transition.intent.url, {