+++ tests/unit/mixins/data-adapter-mixin-test.js
@@ -1,10 +1,8 @@
-import Ember from 'ember';
+import EmberObject from '@ember/object';
 import { describe, beforeEach, it } from 'mocha';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import DataAdapterMixin from 'ember-simple-auth/mixins/data-adapter-mixin';
-
-const { Object: EmberObject } = Ember;
 
 describe('DataAdapterMixin', () => {
   let adapter;
@@ -45,11 +43,23 @@ describe('DataAdapterMixin', () => {
       expect(hash).to.have.ownProperty('beforeSend');
     });
 
-    it('asserts the presence of authorizer', function() {
+    it('asserts `authorize` is overridden', function() {
       adapter.set('authorizer', null);
+
       expect(function() {
         adapter.ajaxOptions();
+        hash.beforeSend();
       }).to.throw(/Assertion Failed/);
+    });
+
+    it('calls `authorize` when request is made', function() {
+      const authorize = sinon.spy();
+      adapter.authorize = authorize;
+      adapter.set('authorizer', null);
+      adapter.ajaxOptions();
+      hash.beforeSend();
+
+      expect(authorize).to.have.been.called;
     });
 
     it('preserves an existing beforeSend hook', function() {
@@ -82,7 +92,7 @@ describe('DataAdapterMixin', () => {
 
       describe('when the authorizer calls the block', function() {
         beforeEach(function() {
-          sinon.stub(sessionService, 'authorize', (authorizer, block) => {
+          sinon.stub(sessionService, 'authorize').callsFake((authorizer, block) => {
             block('header', 'value');
           });
           hash.beforeSend(xhr);
@@ -150,7 +160,7 @@ describe('DataAdapterMixin', () => {
 
     describe('when the authorizer calls the block', function() {
       beforeEach(function() {
-        sinon.stub(sessionService, 'authorize', (authorizer, block) => {
+        sinon.stub(sessionService, 'authorize').callsFake((authorizer, block) => {
           block('X-Authorization-Header', 'an-auth-value');
         });
       });