+++ tests/unit/mixins/unauthenticated-route-mixin-test.js
@@ -1,17 +1,18 @@
-import Ember from 'ember';
+import Mixin from '@ember/object/mixin';
+import RSVP from 'rsvp';
+import Route from '@ember/routing/route';
 import { describe, beforeEach, it } from 'mocha';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import UnauthenticatedRouteMixin from 'ember-simple-auth/mixins/unauthenticated-route-mixin';
 import InternalSession from 'ember-simple-auth/internal-session';
 import EphemeralStore from 'ember-simple-auth/session-stores/ephemeral';
-
-const { Mixin, RSVP, Route } = Ember;
+import createWithContainer from '../../helpers/create-with-container';
 
 describe('UnauthenticatedRouteMixin', () => {
   let route;
   let session;
-  let transition;
+  let containerMock;
 
   describe('#beforeModel', function() {
     beforeEach(function() {
@@ -20,18 +21,19 @@ describe('UnauthenticatedRouteMixin', () => {
           return RSVP.resolve('upstreamReturnValue');
         }
       });
-
       session = InternalSession.create({ store: EphemeralStore.create() });
-      transition = {
-        send() {}
+      containerMock = {
+        lookup: sinon.stub()
       };
 
-      route = Route.extend(MixinImplementingBeforeModel, UnauthenticatedRouteMixin, {
+      containerMock.lookup.withArgs('service:session').returns(session);
+
+      route = createWithContainer(Route.extend(MixinImplementingBeforeModel, UnauthenticatedRouteMixin, {
         // pretend this is never FastBoot
-        _isFastBoot: false,
         // replace actual transitionTo as the router isn't set up etc.
         transitionTo() {}
-      }).create({ session });
+      }), { session }, containerMock);
+
       sinon.spy(route, 'transitionTo');
     });
 
@@ -44,24 +46,24 @@ describe('UnauthenticatedRouteMixin', () => {
         let routeIfAlreadyAuthenticated = 'path/to/route';
         route.set('routeIfAlreadyAuthenticated', routeIfAlreadyAuthenticated);
 
-        route.beforeModel(transition);
+        route.beforeModel();
         expect(route.transitionTo).to.have.been.calledWith(routeIfAlreadyAuthenticated);
       });
 
       it('does not return the upstream promise', function() {
-        expect(route.beforeModel(transition)).to.be.undefined;
+        expect(route.beforeModel()).to.be.undefined;
       });
     });
 
     describe('if the session is not authenticated', function() {
       it('does not call route transitionTo', function() {
-        route.beforeModel(transition);
+        route.beforeModel();
 
         expect(route.transitionTo).to.not.have.been.called;
       });
 
       it('returns the upstream promise', function() {
-        return route.beforeModel(transition).then((result) => {
+        return route.beforeModel().then((result) => {
           expect(result).to.equal('upstreamReturnValue');
         });
       });