+++ tests/unit/services/session-test.js
@@ -1,12 +1,14 @@
-import Ember from 'ember';
+import ObjectProxy from '@ember/object/proxy';
+import Evented from '@ember/object/evented';
+import { next } from '@ember/runloop';
+import { set } from '@ember/object';
+import { registerDeprecationHandler } from '@ember/debug';
 import { describe, beforeEach, it } from 'mocha';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import Session from 'ember-simple-auth/services/session';
 
 import createWithContainer from '../../helpers/create-with-container';
-
-const { ObjectProxy, Evented, run: { next }, set } = Ember;
 
 describe('SessionService', () => {
   let sessionService;
@@ -177,6 +179,23 @@ describe('SessionService', () => {
         expect(() => {
           sessionService.authorize('bad-authorizer', 'block');
         }).to.throw(Error, /No authorizer for factory/);
+      });
+
+      it("shows deprecation warning when 'authorize' is called", function() {
+        let warnings = [];
+        registerDeprecationHandler((message, options, next) => {
+          // in case a deprecation is issued before a test is started
+          if (!warnings) {
+            warnings = [];
+          }
+
+          warnings.push(message);
+          next(message, options);
+        });
+
+        sessionService.authorize('authorizer', 'block');
+        expect(warnings).to.have.length(1);
+        expect(warnings[0]).to.equal("Ember Simple Auth: 'authorize' is deprecated.");
       });
     });